Move test on pre-commit mode
